Add ability to re-generate mocks with "make mocks"
The "mocks" target will take care of installing mockgen and will invoke mockgen
using "go generate". We install a specific version of gomock / mockgen to avoid
having to check-in new source code every time a new version of gomock is
released.

pkg/test/mocks/ovs_mock.go included a MockOVSConfigError struct (implementing
the ovsconfig.Error interface), but it seems that this type was not generated by
mockgen. Additionally, MockOVSConfigError was not really used as a mock in tests
(e.g. no expectations). It seemed more appropriate to replace it with a "dummy"
implementation of ovsconfig.Error, called DummyOVSConfigError (in the
okn/pkg/test package), used specifically for unit tests.