Install iptables in the okn-ubuntu Docker image
The OKN client runs in a Docker container and needs to have access to
iptables. To investigate this issue, it was convenient to do some
refactoring of our iptables code. In particular I believe it is
unpractical to instantiate the IPTables object from the init() function,
since it does not let us log errors (klog is initialized later, in the
main() function) and will lead to a panic in case of an issue.