Mount host's /proc directory as /host/proc in okn-agent container
Without this change the veth interface pair cannot be created and the
CmdAdd fails.

Fixes #2

This commit also re-organizes the ubuntu Dockerfile so that the apt-get
command does not need to run again if only the okn binaries have been
modified.

Finally, this commit ensures that the /var/run/okn/cni.sock does not
exist before we try to bind it to a Unix socket (or we may get an
"address already in use" error). We do that by simply calling os.Remove
before the bind. Without this change one gets an error when deleting the
okn-agent daemon set, then restarting it. There would be other ways of
course to "fix" this issue. For example one could write an additional
initContainer to remove the file. Or we could try to trap SIGTERM /
SIGKILL in the okn-agent process and delete the file then. However, I
think that the solution I chose is both the simplest and the most
effective.