enhance: flatten params in `InstallNodeFlows` and `InstallPodFlows`
The existing InstallNodeFlows and InstallPodFlows use structs to
wrap some parameters. Although it may let the function signature
clean, the invoke side would need more initialization to fill the
struct.Besides it is not an idiomatic way to do like this in the
Go community.

This patch flattens struct parameters in InstallNodeFlows and
InstallPodFlows and updates existed invoking codes.

Resolved #6