Add IPSec tunnel config option and read IKE PSK from an env variable
This commit adds an Agent config option to eanble the IPSec tunnels
and lets Agent read the IKE PSK string from an environment variable:
ANTREA_IPSEC_PSK.

IPSec (ESP) tunnel could be enabled by setting the config option
`enableIPSecTunnel` to "true" in the Agent configuration yaml. IKE
(strongSwan) will be configured to use Preshared Key (PSK) to
authenticate with peers. For this, a K8s Secret should be created to
save the PSK value, and the ANTREA_IPSEC_PSK environment variable
should be added to the Agent container which consumes the PSK key of
the Secret.

A sample PSK Secret is like:
```
apiVersion: v1
kind: Secret
metadata:
  name: antrea-ipsec
type: Opaque
stringData:
  psk: changeme
```