Ensure that Pod connectivity is preserved when okn-agent restarts
It was observed a while ago that Pods could no longer reach each other
after a restart, because of missing OVS flows (the flows go away when
OVS restarts). This patch introduces reconciliation on startup in the
CNI server. The server now gets the current list of Pods from the K8s
apiserver and ensures that the necessary flows exist. If we detect that
the interface store is out-of-date (Pod interface that no longer exists)
we also take care of that. Note that we rely on the interface store
(initialized from the persistent OVSDB) to be able to map each Pod to
its interface.