Add options to configure MTU to okn-agent and CNI configuration
Enable configuration of the MTU to achieve good networking performance
for traffic tunnelled by OVS.

okn-agent needs to be aware of the desired MTU in order to configure the
host gateway's MTU (Node to Pod traffic goes through the gateway before
being tunnelled).

The MTU setting provided to okn-agent will also be used for the Pods'
networking interface unless overwritten by the CNI conf.

If no MTU is provided to okn-agent, we default to 1450 bytes.