Save OKN logs for e2e tests which fail
We invoke "kubectl logs" and "kubectl describe" for every OKN Pod
(across all Nodes) and write the result to disk. This is done for each
test case which fails. If --logs-export-on-success is provided then we
do it for every test case, regardless of whether it fails or not.

By default, the files are written under a temporary directory which is
different for every run. A different existing directory can be selected
using --logs-export-dir. Note that the contents of this directory may be
overwritten.