Cleanup OVS client integration tests
These tests are integration tests (the Makefile was explictly excluding
them for the "test-unit" target) and belong under test/integration.

The tests were not working because the OVSDB UDS socket was an empty
string. I fixed that and switched to using testify for test assertions
for improved readability. I also added a test to make sure that calls to
DeletePort are idempotent.