Update APIServer Storage to take empty interface as argument
The stored objects will be normal structs, not the apis structs that
implement runtime.Object. This patch updates the storage Interface and
its implementation to take empty interface as argument to support it.

Besides, it makes ram.store can take nil genEventFunc so that
NetworkPolicyController can be developed individually without knowing
how to generate events.