Ensure that InstallPodFlows, InstallNodeFlows and InstallServiceFlows are idempotent
Without this change, if one of these functions is called multiple times
with the same key (Node name, Pod container ID, Service name), we issue
an Openflow ADD request and insert the flows in the cache again
(therefore creating duplicates). The result of GetFlowTableStatus would
also be wrong, since flow count is incremented every time ADD is called,
regardless of whether the flow is present.

By making these functions idempotent, we ensure that clients can call
them without having to worry about whether the flows already exist, and
we handle the complexity in the OpenFlow client.