Fix CNI server integration tests
There was a bug in the function that verifies that Pod networking was
setup properly. The code was not using the correct network namespace
when querying the IP addresses assigned to the Pod's interface. On
developers' machines this was going unnoticed because 1) the ifindex
also mapped to an IP interface in the test namesapce, and 2) the test
was only checking for the number of IP addresses assigned to the
interface and not checking whether the IP addresses themselves were
correct.

With this change, we now lookup the IP addresses using the correct
network namespace. The test was also updated to properly verify each IP
address assigned to the Pod's interface.