Use kustomize to generate Antrea manifests
Using kustomize enables us to easily generate different manifests for
different use cases (at the moment development and production /
releases). In addition to that, we no longer put everything into one big
manifest file, but we can place different K8s resources into different
files, which makes maintaining the manifest easier.

kustomize also enables us to easily apply patches to any manifest
(development or release) so that Antrea can be deployed on a Kind
cluster. For example, Kind requires us to use OVS in userspace mode
(netdev datapath), which in turns means that the tun device driver needs
to be available in the antrea-ovs container.

This commit introduces some scripts to make installing / running
kustomize very easy, along with a new make target (make manifest) which
takes care of re-generating the "standard" development manifest, so that
it can be checked-in.

Because one of the motivation for this commit is being able to easily
deploy Antrea in Kind clusters, we also make the necessary changes to
make it possible, and we add some documentation showing how to deploy
Antrea with Kind (see docs/kind.yml).

Fixes #22