Make the cache for Node flows, Pod flows and service flows thread-safe
The assumption is that only calls for different cache keys (host name,
container ID, service name) can happen concurrently. We want to be able
to execute such calls in parallel. It is the responsibility of the
clients of the interface to ensure that calls for the same cache key
cannot happen concurrently. For K8s controllers, this is guaranteed by
the work queue interface. For the CNI server, there is no obvious
guarantee from the CNI specification that concurrent CNI binary
invocations for the same container ID are not possible. Therefore we add
some lightweight logic to the CNI server to ensure that requests for the
same container ID are processed sequentially.

This commit also refactors flow deletion to unify the code between the
different Uninstall* methods.