Use pre-built OVS Docker image
We will now use a pre-built OVS image, hosted in the Antrea Dockerhub repository
as antrea/openvswitch:2.11.1. This image is used as a base image when building
the main Antrea Docker image (antrea-ubuntu), which considerably speeds-up fresh
builds. This commit also documents how to build and push a new image for OVS, in
case we need to use a different OVS version for Antrea.

Fixes #39