Add NetworkPolicy event handler
Add logic to create objects on NetworkPolicy events. This commit
adds handler methods for K8s NetworkPolicy Add/Modified/Delete
events, which lead to the generation of an internal NetworkPolicy
object corresponding to this K8s NetworkPolicy and an
AppliedToGroup object corresponding to the NetworkPolicy
spec.PodSelector. Similarly, it creates an AddressGroup object
for every podSelector/namespaceSelector in the NetworkPolicy
ingress/egress rule.