Fix broken "go mod" because of missing Octant source code
"go mod" was unstable with go 1.12 and broken with go 1.13 because of
some missing test source files in the Octant repository. "go mod tidy"
records test dependencies which means that even generated source files
(e.g. mocks) which are only used for testing have to be checked-in.

This has been patched in the Octant master branch but because we are
using an older version of Octant (v0.8), we are using an Octant fork at
the moment, which includes a back-port of the change to v0.8.

Fixes #81