Always include /opt/cni/bin for searching the IPAM CNI plugin
When Antrea agent runs as a Pod, the IPAM plugin is always installed
in /opt/cni/bin, but kubelet can be configured to use different
paths to search for CNI plugins. This PR always adds /opt/cni/bin to
the CNI plugin searching paths to make sure the IPAM plugin installed
in the agent Pod can be found.