Fix race condition in unit test case TestDeleteRule
The mockReconciler is not thread-safe. The patch fixes it and also
changes to use a separate channel to receive deleted events in case
that there're several Reconcile calls when events are handled in
specific order, which results in multiple notification in "updated"
channel and affect the test of rule deletion.