Also run CI jobs (Go / Kind) on push events
Otherwise we may not detect a situation where a PR is merged and even
though the CI was "green" for the PR, the master branch is broken. An
alternative would be to require PRs to always be up-to-date before
merging but IMO that doesn't scale very well and starts being painful
even with a few PRs a day.

This change will also ensure that the CI badge we display in the README
is accurate.