Create tunnel ports and add OVS flows for IPSec tunnel
For each remote Node, create a separate tunnel port and interface on
the OVS bridge, and set the remote IP and PSK options to the
interface.
Refactore InterfaceStore structs and funcs to add tunnel interface
attributes and group InterfaceConfig fields by interface types.

For packets to a tunnel, change the l3Forwarding table flow to set the
output port to the tunnel interface, and resubmit to the
l2ForwardingOutput table (bypassing the l2ForwardingCalcTable and the
ingress rules tables).
When IPSec is enabled, for every remote Node set the output port to
the Node's tunnel interface, and also add a separate classification
flow in the in port classification table for the tunnel.