Run integration tests as part of CI
This required fixing yet another error in the CNI server test: when
checking for interface deletion, the test was not using the correct
network namespaces / interface names. This was going unnoticed on dev
machines when there was no "eth0" interface. However, the Github CI
worker VM seems to have an eth0 interface. When checking for container
interface deletion, the test would look for an "eth0" interface in the
root namespace (instead of the container namespace) and expect not to
find it. But because of the network configuration for the Github CI
worker VM - i.e. because of the presence of an eth0 interface - the test
was failing (for the wrong reason).

Fixes #179