Fix rule for INVALID packets in conntrackStateTable
Before this change, the flow was matching on ct_state=+new+inv and
dropping all matching packets. According to the OVS documentation:
>If inv is set, only the trk flag is also set.

Which makes sense because an invalid packet is a packet whose state
cannot be determined. Therefore it seems that in its current state, the
rule can never be matched. It should be changed to match on
ct_state=+inv+trk (should be equivalent to ct_state=+inv, but I added
+trk to match the style of other flows).