Add documentation for Antrea's OVS pipeline
Some detailed documentation for the OVS pipeline, including a
description of each table. This is directed at developers and people
trying to troubleshoot issues.

It includes a SVG high-level diagram of the pipeline. We use SVG
directly so it renders better on all screens and to avoid having to
check-in a "large" PNG image that may need to be updated often.

More documentation specific to the Network Policy implementation will
follow later.

Fixes #27