Disable workflows that push to Dockerhub for forks
Such jobs are bound to fail - as they should - when run from a fork
(since the fork does not have permission to push to the dockerhub repo).

It is a bit verbose, and the workflows are still running (i.e. a worker
VM is created) which consummes some compute resources, but I believe the
Github team is working on improving this.

This is the best way I found to do this. I believe Github is working on
a better way to skip workflows.