Refactor kustomize usage
To accomodate for new use cases (ipsec), we refactor how we use
kustomize in hack/generate-manifest.sh. We no longer check-in any
overlay, but instead we create a "chain" of overlays dynamically using
"kustomize edit". Ideally in each overlay of the chain we would take
care of a specific "feature" (e.g. ipsec). However, because of multiple
kustomize shortcomings it is not possible at the moment. Furthermore,
because the Antrea agent ConfigMap is a JSON file (and not a list of
key-value pairs), we can only update the config all-at-once, and not
key-by-key.