Run golangci-lint as part of CI
For now we enable the following linters: misspell, gofmt, deadcode
(changes have been made in the code to fix existing issues reported by
these linters). I picked these linters because the current Antrea code
was already mostly conformant with these and I think the probability of
a false positive is close to 0. More linters can be added to the list
later.

In particular, with this change we will be able to catch typos in
comments before merging a commit.