Fix NetworkPolicy span not updated due to a race condition
In SyncAppliedToGroup, it could happen that some NetworkPolicies
refers it after the method queries the related NetworkPolicies. In this
case these policies won't be enqueued after updating AppliedToGroup's
span, which leaves them in a non consistent state.

This PR fixes it by moving the query after storing the updated
AppliedToGroup, so that both of the NetworkPolicies that referred it
before storing it and the ones after storing it can get the right span.