Use bundle to add multiple Openflow entries in one transaction
Provide an interface on the Bridge to support add multiple Openflow entries
in a single transaction. This function is using Openflow bundle feature.

The function returns nil if all the Openflow entries are realized on the OVS,
otherwise the bundle is discarded and the function returns an error to the
client.