Skip TestPodConnectivityAfterAntreaRestart in Kind CI
TestPodConnectivityAfterAntreaRestart destroys and recreates dataplane
which affects existing connection between components (e.g coredns and
antrea-controller) and kube-apiserver somehow and they don't receive
events stream in some time. Since it only happens on Kind cluster, it
might be related to the dataplane implementation.

Before we can resolve it, skip the test in case of tests depending on
coredns and antrea-controller failing.