hange Pod and tunnel network interface name format
For a Pod interface, use "pod/<Pod-Namespace-name>/<Pod-name>" to be
the key for hashing, and Pod-name for the prefix.
For a Node's tunnel interface, use "node/<Node-name>" to be the key,
and Node-name for the prefix.
No more remove '-' in the prefix, and no more append extra bytes to
the prefix when it is shorter than the length limit.
Also change interfaceCache to use interface key as the map key.