Fix events dropped due to full channel buffer
It's easy to get a number of events buffered in watchers' channel in a
scale cluster. This patch increases the buffer size to 1000 and uses a
timeout mechanism to handle the case that a watcher can't consume
events in time. For each event, it's given 50 ms for blocking send;
watchers that can't finish in time will be terminated to guarantee no
event will be missing and the total processing time will not be
affected by a few unresponsive clients.