Replay flows to OVS after a reconnection
After a reconnection event (cause by a vswitchd restart), the agent must
ensure that all flows are re-installed. We do this by using the flow
information cached in the Antrea OF client, since this change is not
very intrusive and doesn't use up a lot of resources. Because new flows
should not be installed during the reconciliation process, we introduce
a RW lock in the client: regular install / uninstall calls can still
occur in parallel, but reconciliation gets exclusive (write) access.

We include integration tests for the new OF client Reconcile method,
along with one new e2e test which ensures that flows get replayed
correctly after an unplanned OVS restart.

Fixes #274