Update docs/kind.md documentation for macOS support
After some investigation, the method we use on Linux
(hack/kind-linux.sh script) to disable TX checksum offloading for the
veth interface of each Kind Node works just as well on macOS (when using
Docker Desktop) as it does on Linux.

As a result, we update the documentation to indicate that both Linux and
macOS are supported, and we rename the script from kind-linux.sh to
kind-fix-networking.sh.

Fixes #155