Delete stale flows on agent restart (#313)
If the antrea-agent restarts, but not the antrea-ovs container, all
existing OVS flows are preserved in ovs-vswitchd. If some changes occur
while the antrea-agent is down (e.g. a Node leaves the cluster, or
Network Policies are updated), it is possible that some of the preserved
flows are now stale and should be deleted. This is why we introduced in
the past the concept of "round number / id" which we increment every
time the agent restarts (it is persisted in OVSDB on the Node), and we
encode as part of the cookie for each flow we install. The rationale for
that is that it would enable us to identify which flows are stale after
a restart: flows which are still required would receive an updated
cookie, while stale flows would still have the old cookie (with the old
round number) and we would be able to delete them by filtering on the
cookie value.

This commit implements the mechanism described above, with the following
caveat: when the agent is initialized, we start a goroutine which sleeps for
10 seconds before deleting stale flows. This is because we want to wait
until all exisiting flows which are still required receive an update
cookie value. Without this sleep, we would be likely to observe some
"flapping" with flows being deleted and then re-installed, which would
cause issues (connectivity issues, network policies not being enforced,
...). But is 10 seconds a good value? What would be a more
"deterministic" way to decide when to do the deletion operation?

Fixes #311