Create release version of antrea-octant
At the moment we have a Github workflow to upload a tagged Docker image
for Octant + the Octant Antrea plugin, for each release. This is
currently broken because of a typo in the Makefile. In addition to
fixing that typo, this change introduces support for generating a
release antrea-octant.yml manifest that can be uploaded with each
release. Despite the availability of the yaml, users will likely need to
edit the file to match their environment before they can apply it. In
the future we may want to enhance the hack/generate-manifest-octant.sh
script so that user can use a single command to deploy Octant and the
Antrea plugin:

./hack/generate-manifest-octant.sh --image octant-antrea-ubuntu:v0.3.0 \
  --octant-accepted-hosts <X> \
  --octant-listener-addr=<Y> | kubectl apply -f -

Fixes #342