* Preserve OVS datapath flows after an agent restart
Without this patch, when the antrea-ovs container exits gracefully
(e.g. as part of a RollingUpdate for the DaemonSet), datapath flows are
deleted, which breaks existing connections. It takes 10 seconds for
exiting connections to start working again. A recent OVS patch changed
the default behavior of ovs-vswitchd so that datapath flows are no
longer deleted on exit. Because this patch is not yet available in any
released OVS version (and won't be for > 6 months), we apply it manually
when we build our base OVS Docker image.

This is a step towards non-disruptive upgrade. However, there are other
factors in play: 1) when the OVS daemons are stopped, the VXLAN device
(vxlan_sys_4789) goes away, which impacts connectivity between Nodes,
and 2) when ovs-vswitchd restarts, datapath flows are flushed and we
have to wait until the Antrea Agent has replayed the flows.

* Disable new e2e test for Kind

When using the netdev datapath, packet forwarding is done by the
ovs-vswitchd daemon itself, so even existing connections are broken.

* Use arping instead of ping for test

When using ping the test can be flaky: if one of the hosts need to send
an ARP request to refresh its ARP entry while the test is ongoing, odds
are that there is no cached flow in the datapath. If this happens, the
host will mark the ARP entry has "incomplete" and the ping will start
failing.

Fixes #355