Still use flow based tunnel when IPSec encyption is enabled (#359)
Change to use flow based tunnel for IPSec encyption as well, so OVS
flows will be the same no matter IPSec is enabled or not. A separate
tunnel port/interface is still created for each remote Node, which is
required for the OVS IPSec monitor daemon to detect the new tunnel and
create IPSec security policies for the Node using psk and remote_ip
options of the tunnel interface.