Fix agent caching duplicate pod member
When iterating a slice and storing its items' address, must not use the
address of the loop iterator variable as it's the same address taking
different values in each loop iteration, otherwise the stored items
would eventually contain only the last item.
See https://github.com/golang/go/wiki/CommonMistakes#using-reference-to-loop-iterator-variable

This patch fixes it and hardens the unit test to validate the value of
stored items, instead of their keys.