Remove unnecessary periodical resync (#387)
NetworkPolicyController (of antrea-controller) is the only writer of the
antrea network policy storages and it will keep re-enqueuing failed
items until they succeed, so there's no need to have the informers
trigger all items resync, which could introduce considerable workload in
scale cluster. This patch removes it and updates the default
resyncPeriod of informer factory to 12 hours.

NodeRouterController doesn't have heavy workload and its periodical
resync could potentially fix the routes which might be broken by
external so keep it as is.