Add openapi endpoint to antrea-controller apiserver (#395)
Without the openapi endpoint, kube-apiserver will be full of the
error messages:
loading OpenAPI spec for "v1beta1.system.antrea.tanzu.vmware.com" failed with: OpenAPI spec does not exist

This patch generates openapi source file and installs the openapi
handler for antrea-controller apiserver. It's fine that the openapi spec
is for the API group networking.antrea.tanzu.vmware.com, instead of the
registered group system.antrea.tanzu.vmware.com as it is only used to
build an aggregated OpenAPI specs for kube-apiserver.