Reduce network policy calculation log (#398)
There were M * N logs to show whether a Pod matches a NetworkPolicy if
there are M Pods and N NetworkPolicies in the cluster, which generates
log spam and hard to filter useful information in a scale cluster. This
patch changes to log matched item only.

Similarly, AddressGroup and AppliedGroups may have thousands of members,
dumping all of them in one line makes log hard to read. This patch logs
the count of members and span nodes instead.

Besides, this patch unifies the log style.