* Add sonobuoy wrapper script to run Antrea K8s community tests
The wrapper script provides a convenient way to run the subset of K8s
community tests which are relevant to Antrea. The tests are the same as
the ones we use for CI (and maybe the Jenkins jobs could use this script
to run the tests, in order to avoid duplication?).

At the moment the script requires a working cluster, but we can provide
yet-another-script (or extend ci/kind/test-e2e-kind.sh) to take care of
1) creating an appropriate Kind cluster and 2) running the community
tests.

At the moment the script does not support running a locally-modified
version of the upstream tests, but this is something we should
investigate and consider adding in the future.

 * Update README for Jenkins CI tests

The document was not updated when the list of tests was updated in the
job's YAML file.

The document was also updated so that lines wrap at 80 characters.