Cancel OpenFlow message or return error when OVS is disconnected
Add a context in OFSwitch instance, which is used to cancel the OpenFlow
message when the Switch is disconnected. Add timeout in both functions to
send OpenFlow message and wait for bundle reply. The implemention is in
ofnet, bump up version to leverage the fix.
Add integration test to simulate OVS restart case: remove the bridge, and
then add a new bridge with the same name.