Run netpol tests as part of CI
This is good as it adds extra coverage for Antrea, but also ensures that
the prototype netpol tests don't become broken while they are hosted as
part of this repository.

This change also incldues some small fixes in the instructions for
running netpol as a K8s Job.

Because of space limitations on Github CI workers, there was extra
motivation to keep Docker images small:
 * we use a multi-stage build for the netpol image
 * we also use a smaller image for test containers (python:latest is
   quite large, so we build our own based on Alpine).