Always retry to connect the OVS bridge in OpenFlow client after it is disconnected
Use the max retry times in function `OFBridge.Connect`to avoid Agent
initialization is blocking because OVS is not connected. After the
connection is lost, OFClient retry to connect the OVS Bridge until
the new connection is setup, or the OFBridge is deleted.

The fix code is in ofnet, this change is to consume that fix.

Add a test case to check OVS could be re-connected even if it costs more
time to start than the max delay configured in the `Connect` function.