[Netpol] Use parallel probes to verify reachability matrix
This speeds up the time required to run the tests considerably (from 9
minutes to less than 4 minutes). We currently create a goroutine for
each probe, but that may become an issue if the reachability matrix
becomes too large.

We also update the netpol-test image to use ncat (from the nmap
package) instead of the Python SimpleHTTPServer. The Python server
cannot handle concurrent connections which caused the tests to be
flaky. By using ncat, we also make it possible to write network policy
tests for the SCTP protocol.