CI: Wait for Machines to be Actually Deleted in Cleanup Job (#485) (#495)
Due to CAPV issue [1], if Cluster and VSphereCluster are deleted before
Machines, the Machines cannot be deleted. In this patch the Jenkins job
firstly deletes Machines and waits for them to disappear, then it deletes
VSphereCluster and Cluster. This deletion order is the reverse order of
the resource creation. It should minimize the chance of Cluster API
resources getting stuck in deleting state.

[1] https://github.com/kubernetes-sigs/cluster-api-provider-vsphere/issues/761

Co-authored-by: Zhengsheng Zhou <zhengshengz@vmware.com>