Run linters on netpol Go code as part of CI
We also remove the `test-fmt` target since it has been replaced by
`golangci`. It is still possible to use `make fmt` to apply `gofmt`
changes on the entire code base.

Because `golangci-lint` can only be used in the scope of a single module
(apparently), we need a separate `golangci` make target in
hack/netpol. This is probably better anyway as it helps keep the Antrea
core code and the netpol code separate, which is why they are different
modules in the first place.