Add e2e test for upgrade scenario
The new Go test (TestUpgrade) is meant to be run through the wrapper
script ./ci/kind/test-upgrade-antrea.sh.

At the moment we test upgrade from versions 0.3.0, 0.4.0 and 0.4.1 as
part of CI, but we can modify that set as we go through Antrea
releases. We also run the test for every PR against master, which may be
a bit excessive.

Fixes #511