Start documentation for antctl
Add a new document under docs/ with information on how to install
(instructions valid starting with release 0.5.0) and use antctl. This is
user-facing documentation and does not include information about the
antctl implementation.

We also do not include a detailed list of commands at the moment. This
list can come later if needed (can be hard to keep up-to-date), and we
can keep adding information about specific commands in
troubleshooting.md for specific debugging scenarios.

Fixes #337