Fix Antrea EKS manifest
The PR adding the antrea-eks.yml manifest was merged without being
rebased first, so the new ServiceAccount used by antctl was omitted from
it.

This commit also fixes a typo in the ServiceAccount's name.
The PR adding the new ServiceAccount was merged before the PR adding the
antrea-eks.yml manifest