Do not use Github API to check tag
There seems to be some rate limiting which creates issues when running
an unauthenticated request as part of a Github workflow.

Furthermore, using "grep -q" with "curl" sometimes lead to the "(23)
Failed writing body" error being displayed when grep closes the pipe
early. Not sure exactly why since we run curl in silent mode, but I have
seen it in the workflow logs.