Check whether an object matches the watcher before generating init event (#543)
It's observed in scale tests that antrea-controller had memory usage
peak when agents were connecting to it, which is quite greater than the
usage after agents connected it. For example, in a scale of 100k Pods,
75k NetworkPolicies, 25k Namespaces, 1k antrea-agents, it would consume
8~9GB memory at peak and lead to OOM sometimes while it only consumed
1.3GB memory when there were no agents connection and 1.5GB after 1k
Nodes connected it.

The memory should be consumed by a process in store.Watch where it lists
all items (pointers) of a kind of resource, converts them to
InternalEvent, sends them to watchers to further process. This means
that every object will be converted once regardless of whether the
watcher is interested in it or not, while the conversion is an expensive
operation in memory usage and normally a watcher won't be interested in
most objects, for example, in the above scale each agent only cares
hundreds out of 75k NetworkPolicies.

This patch optimizes it by filtering the object before converting it. As
the filter is a very cheap hash query, it can save the memory and CPU
usage a lot. For example, in a local scale test with 80 Pods, 30k
NetworkPolicies, 10k Namespaces, 50 antrea-agents, it would consume
1.1~1.2GB memory at peak. With this patch, it would consume only 310~MB
at peak.