Improve antrea-controller NetworkPolicy computation performance
In a scale of 25k Namespaces, 100k Pods, 75k NetworkPolicies, It took
~20 minutes to finish the initial computation, which could leave the
cluster in a degraded state for a long time once antrea-controller is
restarted.

It's profiled that the bottleneck is the Pod event handlers: addPod. For
instance, in the above scale, addNetworkPolicy can finish all tasks in
9.42s, addNamespace can finish in 910ms, but addPod needs 17min. So it
is actually addPod keeping producing new tasks for the workers to
process. It's slow because addPod listed all AddressGroups and
AppliedToGroups and checked whether they could be affected by this pod,
leading to O(N * M) time complexity, where N is number of pods, M is
number of groups. Further more, as multiple Pods can affect same
AddressGroups or AppliedToGroups, if it produces very slow but the
workers consumes very fast, there would be many repeated computation,
for instance: podA added -> sync groupM -> podB added -> sync groupM ->
podC added -> sync groupM, where groupM is synced 3 times, wasting CPU a
lot, making the whole processing worse.

Given the fact that a Pod can only possibly affect AddressGroups and
AppliedToGroups that are namespace scoped and targeting for this
namespace, and that are cluster scoped. This patch builds a namespace
index for AddressGroups and AppliedToGroups. When it processes a Pod
added event, it no longer lists all groups but get a small portion of
groups by index, ideally the while time complexity is reduced to O(N),
so it is able to produce all tasks for workers very fast. Besides, the
workers will only process each group once or a few times.

With this optimization, the execution time of the above scale is reduced
from ~20 minutes to 6 seconds.

The patch also adds a performance test with a large scale fake input to
measure the metrics automatically.