[Netpol] Sleep between successive tests
Because the same Pods are re-used to validate NetworkPolicy
implementation, there is virtually no delay between successive
tests. This means that Antrea does not always have time to properly
clean-up stale NetworkPolicies and a test can fail because of the
NetworkPolicy from the previous test. By adding a short delay (2s) every
time we update or delete a NetworkPolicy, we can bypass the issue. This
change only increases testing time by < 20% (3min -> 3min30).

This commit also changes the verbosity of some log messages to "TRACE"
to make the logs easier to navigate.

Fixes #544